Renamed Controllers dir name to Controller, makes it more consistent with the setup for a symfony bundle
